# SWAGGER-NVSL-APIS.yaml
openapi: 3.0.3
info:
  title: NVSL API
  version: 1.0.0
  description: |
    Documentação consolidada das APIs do NVSL.
    Base URL padrão: `http://localhost:8081/api`.

    Rotas **web** na raiz da aplicação (`http://localhost:8081`, cadastro MGI):
    - `GET /` — status da API
    - `GET /redirect-gov` — callback OAuth2/OIDC (redirect_uri)
    - `GET /logout` — redireciona ao login do frontend (`?from=logout`); revogação do token: `POST /api/auth/logout`
servers:
  - url: http://localhost:8081/api
    description: Ambiente local (Docker)
...
paths:
...
  /auth/redirect:
    get:
      tags: [Autenticação]
      summary: Gera URL de redirecionamento GOV.BR
      responses:
        "200":
          description: URL para iniciar login GOV.BR
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
        "422":
          description: Configuração GOV.BR inválida
  /auth/callback:
    get:
      tags: [Autenticação]
      summary: Callback do GOV.BR (redireciona para frontend)
      description: |
        **Implementação real:** `GET http://localhost:8081/redirect-gov` (rota web, fora de `/api`).
        Mantido aqui apenas como referência do fluxo; o `redirect_uri` cadastrado no MGI deve apontar para `/redirect-gov`.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
        - name: error_description
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redireciona para frontend (/login#...)
  /auth/exchange:
    post:
      tags: [Autenticação]
      summary: Troca login code por token Sanctum
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthExchangeRequest"
      responses:
        "200":
          description: Token e usuário retornados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthExchangeResponse"
        "422":
          description: Código inválido/expirado
  /auth/token-de-teste:
    post:
      tags: [Autenticação]
      summary: Gera token de teste por perfil (somente local/testing)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenTesteRequest"
      responses:
        "200":
          description: Token de teste
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenTesteResponse"
        "403":
          description: Endpoint indisponível no ambiente atual
        "404":
          description: Usuário de teste não encontrado
  /auth/logout:
    post:
      tags: [Autenticação]
      summary: Logout do usuário autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout efetuado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: Não autenticado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    UserSafe:
      type: object
      properties:
        id: { type: integer }
        govbr_sub: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        role: { type: string }
        esfera_atuacao: { type: string }
        uf_lotacao: { type: string, nullable: true }
        municipio_lotacao: { type: string, nullable: true }
    TokenTesteRequest:
      type: object
      properties:
        perfil:
          type: string
          enum: [federal, estadual, municipal]
          default: federal
    TokenTesteResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/UserSafe"
    AuthExchangeRequest:
      type: object
      required: [code]
      properties:
        code: { type: string }
    AuthExchangeResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/UserSafe"
...